+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ﻿# Changelog
 
+## 1.1.0
+- added new class RawBase64String
+- moved methods rawFromBytes, rawFromText, rawToBytes and rawToText to the new class RawBase64String and
+  renamed them as encodeBytes, encodeText, decodeAsBytes and decodeAsText
+- added methods encode and decode to RawBase64String
+- renamed methods fromBytes, fromText, toBytes and toText of the Base64String class as encodeBytes, encodeText, decodeAsBytes and decodeAsText
+- added methods encode and decode to Base64String
+
 ## 1.0.0
 
 ## 0.8.0